requirements.txt
requests
pyyaml
paho-mqtt